pyproject.toml
[project]
name = "c2s-dspy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=2.6.27",
    "openai>=1.85.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
